[rust] Move tests out of lib.rs to dedicated file
