+++ crates/petname/src/lib.rs
@@ -68,76 +68,3 @@ mod ffi {
         fn make_name(petname: &mut Petname) -> String;
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use std::collections::HashSet;
-
-    fn split_name<'a>(name: &'a str, separator: &str) -> Vec<&'a str> {
-        name.split(separator).collect()
-    }
-
-    #[test]
-    fn generates_the_requested_num_words() {
-        let mut gen1 = new_petname(1, "-");
-        let mut gen2 = new_petname(2, "-");
-        let mut gen3 = new_petname(3, "-");
-
-        let one_word_name = make_name(&mut gen1);
-        let tokens = split_name(&one_word_name, "-");
-        assert_eq!(tokens.len(), 1);
-
-        let two_word_name = make_name(&mut gen2);
-        let tokens = split_name(&two_word_name, "-");
-        assert_eq!(tokens.len(), 2);
-
-        let three_word_name = make_name(&mut gen3);
-        let tokens = split_name(&three_word_name, "-");
-        assert_eq!(tokens.len(), 3);
-    }
-
-    #[test]
-    fn uses_custom_separator() {
-        let mut name_generator = new_petname(3, "_");
-        let name = make_name(&mut name_generator);
-        let tokens = split_name(&name, "_");
-        assert_eq!(tokens.len(), 3);
-        assert!(name.contains("_"));
-    }
-
-    #[test]
-    fn generates_two_tokens_by_default() {
-        let mut name_generator = new_petname(2, "-");
-        let name = make_name(&mut name_generator);
-        let tokens = split_name(&name, "-");
-        assert_eq!(tokens.len(), 2);
-
-        // Each token should be unique
-        let unique_tokens: HashSet<_> = tokens.iter().collect();
-        assert_eq!(unique_tokens.len(), tokens.len());
-    }
-
-    #[test]
-    fn can_generate_at_least_hundred_unique_names() {
-        let mut name_generator = new_petname(3, "-");
-        let mut name_set = HashSet::new();
-        let expected_num_unique_names = 100;
-
-        // Generate many names to test uniqueness
-        for _ in 0..(10 * expected_num_unique_names) {
-            let name = make_name(&mut name_generator);
-            name_set.insert(name);
-        }
-
-        assert!(name_set.len() >= expected_num_unique_names);
-    }
-
-    #[test]
-    fn ffi_integration_test() {
-        let mut petname = new_petname(2, "-");
-        let name = make_name(&mut petname);
-        assert!(name.contains("-"));
-        assert!(!name.is_empty());
-    }
-}
